Extract ShareButtons to components
Fix svg icon missing from block in editor
