+++ classes/class-loader.php
@@ -354,6 +354,8 @@ final class Loader {
 		// Variables reflected from PHP to JS.
 		$reflection_vars = [
 			'dateFormat'                => get_option( 'date_format' ),
+			'siteUrl'                   => site_url(),
+			'themeUrl'                  => get_template_directory_uri(),
 			'enable_analytical_cookies' => $option_values['enable_analytical_cookies'] ?? '',
 		];
 		wp_localize_script( 'planet4-blocks-editor-script', 'p4bk_vars', $reflection_vars );
